Moved test to test_system_group_negative as per @lekha's suggestion
